+++ src/rs/lib.rs
@@ -57,7 +57,6 @@ mod tests {
 
     #[test]
     fn w0_000_map() {
-        let mut map = MapFile::new();
-        map.read_map_file(&PathBuf::from("tests/maps/w0_000.map"));
+        let _ = MapFile::new_from_map_file(&PathBuf::from("tests/maps/w0_000.map"));
     }
 }